Fix incorrect comparisons
1. Fix some logical operators
2. The comparison uint16_t != -1 is always true. Should use an int to
save the return value of a function that returns an int.

+++ examples/platform-specific/zoul/test-grove-light-sensor.c
@@ -61,7 +61,7 @@ PROCESS_THREAD(remote_grove_light_process, ev, data)
 {
   PROCESS_BEGIN();
 
-  uint16_t ldr;
+  int ldr;
 
   /* Use pin number not mask, for example if using the PA5 pin then use 5 */
   adc_sensors.configure(ANALOG_GROVE_LIGHT, 5);
