+++ examples/platform-specific/zoul/test-rotation-sensor.c
@@ -61,7 +61,7 @@ PROCESS_THREAD(remote_rotation_process, ev, data)
 {
   PROCESS_BEGIN();
 
-  uint16_t rotation;
+  int rotation;
 
   /* Use pin number not mask, for example if using the PA5 pin then use 5 */
   adc_sensors.configure(ANALOG_PHIDGET_ROTATION_1109, 5);
